LPS-75989 Merge my-account with users-admin-web. This must happen my-account has way too many dependencies on internal packages by coping all the JSPs from users-admin-web.
