--- modules/apps/foundation/my-account/my-account-web/bnd.bnd
@@ -1,13 +1,0 @@
-Bundle-Name: Liferay My Account Web
-Bundle-SymbolicName: com.liferay.my.account.web
-Bundle-Version: 1.0.13
-Import-Package:\
-	javax.net.ssl,\
-	\
-	*
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title:
--dsannotations-options: inherit
--includeresource:\
-	content=../../users-admin/users-admin-web/src/main/resources/content,\
-	META-INF/resources=../../users-admin/users-admin-web/src/main/resources/META-INF/resources
