--- modules/apps/foundation/my-account/my-account-web/build.gradle
@@ -1,24 +1,0 @@
-task copyCompiledJSP(type: Copy)
-
-copyCompiledJSP {
-	from {
-		Project usersAdminWebProject = project(":apps:foundation:users-admin:users-admin-web")
-
-		usersAdminWebProject.compileJSP.destinationDir
-	}
-
-	into {
-		compileJSP.destinationDir
-	}
-}
-
-dependencies {
-	provided group: "com.liferay", name: "com.liferay.application.list.api", version: "2.0.0"
-	provided group: "com.liferay", name: "com.liferay.portal.upgrade", version: "2.0.0"
-	provided group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "2.0.0"
-	provided group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.4.0"
-	provided group: "javax.portlet", name: "portlet-api", version: "2.0"
-	provided group: "javax.servlet", name: "javax.servlet-api", version: "3.0.1"
-	provided group: "org.osgi", name: "org.osgi.service.component.annotations", version: "1.3.0"
-	provided project(":apps:foundation:users-admin:users-admin-api")
-}
