--- modules/apps/foundation/my-account/my-account-web/src/main/resources/portlet.properties
@@ -1,7 +1,0 @@
-include-and-override=portlet-ext.properties
-
-#
-# Input a list of comma delimited resource action configurations that will be
-# read from the class path.
-#
-resource.actions.configs=resource-actions/default.xml
